api/requirements.txt
flask==2.3.2
Flask-JWT-Extended==4.4.0
Jinja2==3.1.2
mysql-connector-python==8.1.0
pandas==1.3.3
bcrypt==4.0.1
PyJWT==2.3.0
python-dotenv==1.0.0
uWSGI==2.0.20
Werkzeug==2.3.6